backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
sqlalchemy==2.0.23
python-dotenv==1.0.0
requests==2.31.0
openai==1.3.7
scikit-learn==1.3.2
numpy==1.26.2
pandas==2.1.4
aiohttp==3.9.1
beautifulsoup4==4.12.2
lxml==4.9.3
schedule==1.2.0
pytest==7.4.3
